x-plat run-jest command
