+++ libs/jestutils/bin/run-jest-nix.cjs
@@ -1,0 +1,3 @@
+#!/usr/bin/env -S node --experimental-vm-modules --no-warnings
+
+require("./run-jest-common.cjs")
